# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sign Takeoff Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadResponse {
  jobId: string;
  fileCount: number;
  message: string;
}

export interface ProcessResponse {
  success: boolean;
  message: string;
  extractedCount?: number;
}

export type JobSummaryStatus =
  (typeof JobSummaryStatus)[keyof typeof JobSummaryStatus];

export const JobSummaryStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface JobSummary {
  id: string;
  status: JobSummaryStatus;
  fileCount: number;
  error?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface JobFile {
  id: string;
  originalName: string;
  pageCount?: number | null;
  createdAt: string;
}

export interface ExtractedSign {
  id: string;
  jobFileId?: string | null;
  sheetNumber?: string | null;
  detailReference?: string | null;
  signType?: string | null;
  signIdentifier?: string | null;
  quantity?: number | null;
  location?: string | null;
  dimensions?: string | null;
  mountingType?: string | null;
  finishColor?: string | null;
  illumination?: string | null;
  materials?: string | null;
  messageContent?: string | null;
  notes?: string | null;
  /**
   * @minimum 0
   * @maximum 1
   */
  confidenceScore: number;
  reviewFlag: boolean;
  createdAt: string;
}

export interface JobDetails {
  job: JobSummary;
  files: JobFile[];
  extractedSigns: ExtractedSign[];
  totalSigns: number;
  flaggedCount: number;
  highConfidenceCount: number;
}

export interface ErrorResponse {
  error: string;
  details?: string | null;
}

export type UploadFilesBody = {
  files: Blob[];
};

export type ListJobs200 = {
  jobs: JobSummary[];
};