# lib/api-zod/src/generated/types/jobFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sign Takeoff Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface JobFile {
  id: string;
  originalName: string;
  pageCount?: number | null;
  createdAt: Date;
}