# lib/api-zod/src/generated/types/jobSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sign Takeoff Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { JobSummaryStatus } from "./jobSummaryStatus";

export interface JobSummary {
  id: string;
  status: JobSummaryStatus;
  fileCount: number;
  error?: string | null;
  createdAt: Date;
  updatedAt: Date;
}