# lib/api-zod/src/generated/types/processResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sign Takeoff Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ProcessResponse {
  success: boolean;
  message: string;
  extractedCount?: number;
}