# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sign Takeoff Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResponse {
  jobId: string;
  fileCount: number;
  message: string;
}